Improve Nginx config file generation

+++ proxy/nginx.sh
@@ -1,4 +1,4 @@
-SMRT_ROOT="" # path to "SMRT_ROOT" directory
+SMRT_ROOT="/opt/pacbio/smrtlink" # optional variable to help find the certificate (see README.md)
 PROXY_HOST=""
 PROXY_DIR=$(dirname $(pwd)/$0)
 CERT=$SMRT_ROOT/userdata/config/security/pb-smrtlink-default.crt
@@ -6,26 +6,42 @@ CERT_KEY=$SMRT_ROOT/userdata/config/security/pb-smrtlink-default.key
 SL_PORT=8243
 PROXY_PORT=8244
 EDWARD_PORT=9093
+SL_USER=""
 
-[ -z "$PROXY_HOST" ] && echo "Please insert value of PROXY_HOST in $0." && exit 1
-[ -z "$SMRT_ROOT" ] && echo "Please insert value of SMRT_ROOT in $0." && exit 1
+# check that USER is SL_USER
+if [ -z $SL_USER ] && echo "Please insert value of SL_USER in $0 (see README.md)." && exit 1
+if [ $USER != $SL_USER ] && echo "Please run this script as $SL_USER (see README.md)." && exit 1
 
 # if the index.html file does not exist, generate it from template
 if [ ! -f $PROXY_DIR/index.html ]; then
-    cp $PROXY_DIR/template.html $PROXY_DIR/index.html
-    sed -i "s|PROXY_HOST|$PROXY_HOST|g" $PROXY_DIR/index.html
-    sed -i "s|PROXY_PORT|$PROXY_PORT|g" $PROXY_DIR/index.html
+    if [ -z $PROXY_HOST ]; then
+        echo "Warning: Please insert value of PROXY_HOST in $0 (see README.md)."
+    else
+        cp $PROXY_DIR/template.html $PROXY_DIR/index.html
+        sed -i "s|PROXY_HOST|$PROXY_HOST|g" $PROXY_DIR/index.html
+        sed -i "s|PROXY_PORT|$PROXY_PORT|g" $PROXY_DIR/index.html
+        echo "Created index.html file."
+    fi
 fi
 
 # if nginx.conf file does not exist, generate it from template
 if [ ! -f $PROXY_DIR/nginx.conf ]; then
-    cp $PROXY_DIR/template.conf $PROXY_DIR/nginx.conf
-    sed -i "s|CERT|$CERT|g" $PROXY_DIR/nginx.conf
-    sed -i "s|SL_PORT|$SL_PORT|g" $PROXY_DIR/nginx.conf
-    sed -i "s|CERT_KEY|$CERT_KEY|g" $PROXY_DIR/nginx.conf
-    sed -i "s|PROXY_DIR|$PROXY_DIR|g" $PROXY_DIR/nginx.conf
-    sed -i "s|PROXY_PORT|$PROXY_PORT|g" $PROXY_DIR/nginx.conf
-    sed -i "s|EDWARD_PORT|$EDWARD_PORT|g" $PROXY_DIR/nginx.conf
+    if [ -f $CERT ]; then
+        if [ -f $CERT_KEY ]; then
+            cp $PROXY_DIR/template.conf $PROXY_DIR/nginx.conf
+            sed -i "s|CERT|$CERT|g" $PROXY_DIR/nginx.conf
+            sed -i "s|SL_PORT|$SL_PORT|g" $PROXY_DIR/nginx.conf
+            sed -i "s|CERT_KEY|$CERT_KEY|g" $PROXY_DIR/nginx.conf
+            sed -i "s|PROXY_DIR|$PROXY_DIR|g" $PROXY_DIR/nginx.conf
+            sed -i "s|PROXY_PORT|$PROXY_PORT|g" $PROXY_DIR/nginx.conf
+            sed -i "s|EDWARD_PORT|$EDWARD_PORT|g" $PROXY_DIR/nginx.conf
+            echo "Created nginx.conf file."
+        else
+            echo "Certificate key not found at $CERT_KEY (see README.md)."
+        fi
+    else
+        echo "Certificate not found at $CERT (see README.md)."
+    fi
 fi
 
 # run nginx
